requirements.txt
gym-retro >= 0.8.0
torch >= 1.5.1
torchvision >= 0.6.0
pillow == 6.2.1
